requirements.txt
Django==3.0.6
django-cors-headers==3.4.0
djangorestframework==3.11.0
djoser==2.0.5

